+++ csv_detective/formats/year.py
@@ -10,15 +10,16 @@ labels = {
 }
 
 
-def _is(val):
-    try:
-        val = int(val)
-    except ValueError:
-        return False
+def _is(val: str | int) -> bool:
+    if not isinstance(val, int):
+        try:
+            val = int(val)
+        except ValueError:
+            return False
     return (1800 <= val) and (val <= 2100)
 
 
 _test_values = {
-    True: ["2015"],
-    False: ["20166", "123"],
+    True: ["2015", 2020],
+    False: ["20166", "123", "2020.5", 156],
 }