+++ csv_detective/output/__init__.py
@@ -3,6 +3,7 @@ import os
 from typing import Iterator
 
 import pandas as pd
+import pyarrow.parquet as pq
 
 from csv_detective.output.dataframe import cast_df_chunks
 from csv_detective.output.profile import create_profile
@@ -11,7 +12,7 @@ from csv_detective.utils import is_url, sanitize_for_json
 
 
 def generate_output(
-    table: pd.DataFrame,
+    table: pd.DataFrame | pq.ParquetFile,
     analysis: dict,
     file_path: str,
     num_rows: int = 500,
@@ -21,6 +22,7 @@ def generate_output(
     output_schema: bool = False,
     output_df: bool = False,
     cast_json: bool = True,
+    na_values: list[str] | None = None,
     verbose: bool = False,
     _col_values: dict[str, pd.Series] | None = None,
 ) -> dict | tuple[dict, Iterator[pd.DataFrame]]:
@@ -64,6 +66,7 @@ def generate_output(
             analysis=analysis,
             file_path=file_path,
             cast_json=cast_json,
+            na_values=na_values,
             verbose=verbose,
         )
     return analysis