+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.12.0 (2026-07-08)
+
+- Add unique values list for (multi)categorical columns ([#250](https://github.com/datagouv/csv-detective/pull/250))
+- Allow to analyse parquet files ([#253](https://github.com/datagouv/csv-detective/pull/253))
+- Allow to pass additional NaN values ([#255](https://github.com/datagouv/csv-detective/pull/255))
+- chore: add related test
+- chore: add test value
+- chore: bump urrlib3 ([#251](https://github.com/datagouv/csv-detective/pull/251))
+- chore: lint
+- chore/use-uv-publish-token ([#256](https://github.com/datagouv/csv-detective/pull/256))
+- clean: fix linting errors
+- feat: only run detection on unique values ([#241](https://github.com/datagouv/csv-detective/pull/241))
+- Fix datetime RFC 822 regex ([#244](https://github.com/datagouv/csv-detective/pull/244))
+- Fix unique values for too complex columns ([#257](https://github.com/datagouv/csv-detective/pull/257))
+- fix: bug on empty columns ([#246](https://github.com/datagouv/csv-detective/pull/246))
+- fix: digits
+- fix: digits
+- fix: fail validation if crash while loading chunks
+- fix(ci): fix lint check in CI
+- Rename `booleen` format into `bool` ([#252](https://github.com/datagouv/csv-detective/pull/252))
+- Update int.py
+- Update siren.py
+- Update siret.py
+- Use iterators in `any` and `all` ([#243](https://github.com/datagouv/csv-detective/pull/243))
+
+
 ## 0.11.2 (2026-04-03)
 
 - feat: do not run any tests on empty columns ([#240](https://github.com/etalab/csv-detective/pull/240))